chore: more refactoring
