+++ src/components/lava-coder/use-cesium-viewer.ts
@@ -1,13 +1,6 @@
-import {
-  Cartesian3, CesiumWidget, createWorldTerrainAsync, Math as CSMath, ImageryLayer, Ion, KmlDataSource, Rectangle,
-  SingleTileImageryProvider, TerrainProvider
-} from "@cesium/engine";
-import { autorun } from "mobx";
+import { Cartesian3, CesiumWidget, Math as CSMath, Ion } from "@cesium/engine";
 import { useEffect, useRef, useState } from "react";
-import hazardZonesKml from "../../assets/Volcano_Lava_Flow_Hazard_Zones.kml";
-import { lavaElevations, lavaSimulation } from "../../stores/lava-simulation-store";
-import { maxLat, maxLong, minLat, minLong } from "./lava-constants";
-import { visualizeLava } from "./visualize-lava";
+import { useTerrainProvider } from "./use-terrain-provider";
 
 import "@cesium/engine/Source/Widget/CesiumWidget.css";
 
@@ -21,7 +14,6 @@ export function useCesiumViewer(container: Element | null) {
   const viewer = useRef<CesiumWidget | null>(null);
   const [ , forceRefresh] = useState(false);
   const terrainProvider = useTerrainProvider();
-  const lavaLayerRef = useRef<ImageryLayer | null>(null);
 
   useEffect(() => {
     if (container && terrainProvider && !viewer.current) {
@@ -44,27 +36,6 @@ export function useCesiumViewer(container: Element | null) {
     }
     return () => viewer.current?.destroy();
   }, [container, terrainProvider]);
-  
-  // Update the lava display
-  useEffect(() => {
-    return autorun(() => {
-      const { coveredCells, raster } = lavaSimulation;
-
-      if (!coveredCells || !lavaElevations || !raster || !widget.current) return;
-
-      const oldLayer = oldLavaLayerRef.current;
-      oldLavaLayerRef.current = lavaLayerRef.current;
-
-      const url = visualizeLava(raster, lavaElevations);
-      lavaLayerRef.current = ImageryLayer.fromProviderAsync(
-        SingleTileImageryProvider.fromUrl(url, {
-          rectangle: Rectangle.fromDegrees(minLong, minLat, maxLong, maxLat)
-        })
-      );
-      if (lavaLayerRef.current) widget.current.imageryLayers.add(lavaLayerRef.current);
-      if (oldLayer) widget.current.imageryLayers.remove(oldLayer, true);
-    });
-  }, []);
 
   return viewer.current;
 }
