+++ src/components/lava-coder/use-hazard-zones.ts
@@ -1,8 +1,8 @@
-import { KmlDataSource } from "@cesium/engine";
+import { CesiumWidget, KmlDataSource } from "@cesium/engine";
 import { useEffect, useState } from "react";
 import hazardZonesKml from "../../assets/Volcano_Lava_Flow_Hazard_Zones.kml";
 
-export function useHazardZones() {
+export function useHazardZones(viewer: CesiumWidget | null, showHazardZones: boolean, verticalExaggeration: number) {
   const [hazardZones, setHazardZones] = useState<KmlDataSource | null>(null);
 
   useEffect(() => {
@@ -14,6 +14,22 @@ export function useHazardZones() {
       console.error("Failed to load KML file:", error);
     });
   }, []);
+
+  useEffect(() => {
+    if (hazardZones) {
+      hazardZones.show = showHazardZones;
+    }
+  }, [hazardZones, showHazardZones]);
+
+  useEffect(() => {
+    if (viewer) {
+      // update hazard zones overlay when vertical exaggeration is changed
+      viewer.dataSources.removeAll();
+      if (hazardZones) {
+        viewer.dataSources.add(hazardZones);
+      }
+    }
+  }, [hazardZones, verticalExaggeration, viewer]);
 
   return hazardZones;
 }
